Temp

+++ index.js
@@ -817,6 +817,8 @@ client.on('message', (msgObj) => {
     let content = msgObj.content;
     const authorId = author.id;
 
+    if (guild.id !== '166601083584643072') return;
+
     if (content.substring(content.length - 5) === ' -del' && authorId === vaebId) {
         msgObj.delete();
         content = content.substring(0, content.length - 5);
